Update README.md
In the LDA projection, the classes are more clearly separated compared to PCA, indicating that LDA captures the discriminative information better for classification. On the other hand, PCA focuses on capturing the overall variance in the
data and is not as class-specific as LDA.
LDA cannot directly project the data into three-dimensional space. If the number of unique classes (c) is greater than
2, LDA can project the data into a two-dimensional space while maximizing the separation between classes. However,
projecting in three-dimensional space would require at least 4(c = 4) dimensions