Add files via upload
These are handwritten digits of 0,1, and 2("MNIST digits0-1-2.csv")
Projected data in 2D LDA and 2D PCA
In the LDA projection, the classes are more clearly separated compared to PCA, indicating that LDA captures the discriminative information better for classification. On the other hand, PCA focuses on capturing the overall variance in the
data and is not as class-specific as LDA.
LDA cannot directly project the data into three-dimensional space. If the number of unique classes (c) is greater than
2, LDA can project the data into a two-dimensional space while maximizing the separation between classes. However,
projecting in three-dimensional space would require at least 4(c = 4) dimensions